+++ app/api/depends.py
@@ -1,0 +1,10 @@
+from typing import Generator
+
+from sqlmodel import Session
+
+from app.models.database import engine
+
+
+def get_db() -> Generator[Session, None, None]:
+    with Session(engine) as session:
+        yield session
