+++ app/api/routes/health.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+from fastapi.responses import JSONResponse
+
+
+router = APIRouter()
+
+
+@router.get("/healthz")
+def healthz() -> JSONResponse:
+    return JSONResponse({"status": "ok"})
